+++ tests/unit_tests/test_runner.py
@@ -62,7 +62,7 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     other_config = Config(tiny_config)
     other_config.n_layers = 50  # lol
     other_config.n_beams = 12
-    other_config.max_iters = 2
+    other_config.cosine_schedule_period_iters = 2
     with torch.device("meta"):
         # Now load the weights into a new model
         # The device should be meta for all the weights.
@@ -72,7 +72,7 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     obs_layers = runner.model.encoder.transformer_encoder.num_layers
     assert obs_layers == 1  # Match the original arch.
     assert runner.model.n_beams == 12  # Match the config
-    assert runner.model.max_iters == 2  # Match the config
+    assert runner.model.cosine_schedule_period_iters == 2  # Match the config
     assert next(runner.model.parameters()).device == torch.device("meta")
 
     # If the Trainer correctly moves the weights to the accelerator,
@@ -97,6 +97,34 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     # Should work:
     with ModelRunner(config=config, model_filename=str(ckpt)) as runner:
         runner.evaluate([mgf_small])
+
+
+def test_save_and_load_weights_deprecated(tmp_path, mgf_small, tiny_config):
+    """Test saving and loading weights with deprecated config options."""
+    config = Config(tiny_config)
+    config.max_epochs = 1
+    config.cosine_schedule_period_iters = 5
+    ckpt = tmp_path / "test.ckpt"
+
+    with ModelRunner(config=config) as runner:
+        runner.train([mgf_small], [mgf_small])
+        runner.trainer.save_checkpoint(ckpt)
+
+    # Replace the new config option with the deprecated one.
+    ckpt_data = torch.load(ckpt)
+    ckpt_data["hyper_parameters"]["max_iters"] = 5
+    del ckpt_data["hyper_parameters"]["cosine_schedule_period_iters"]
+    torch.save(ckpt_data, str(ckpt))
+
+    # Inference.
+    with ModelRunner(config=config, model_filename=str(ckpt)) as runner:
+        runner.initialize_model(train=False)
+        assert runner.model.cosine_schedule_period_iters == 5
+    # Fine-tuning.
+    with ModelRunner(config=config, model_filename=str(ckpt)) as runner:
+        with pytest.warns(DeprecationWarning):
+            runner.train([mgf_small], [mgf_small])
+            assert "max_iters" not in runner.model.opt_kwargs
 
 
 def test_calculate_precision(tmp_path, mgf_small, tiny_config):